% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eucl_dist.R
\name{eucl_dist}
\alias{eucl_dist}
\title{eucl_dist}
\usage{
eucl_dist(P1, P2)
}
\arguments{
\item{P1}{x and y coordinates of the first point}

\item{P2}{x and y coordinates of the first point}
}
\description{
A function to calculate the euclidian distance
}
\examples{
P1<-c(0,0)
P2<-c(1,2)
eucl_dist(P1,P2)
}
